pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ministral-fine-tuning"
version = "0.1.0"
description = "LoRA fine-tuning scaffold for Ministral 3 3B Base"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.5.2",
    "bitsandbytes>=0.46.0",
    "huggingface-hub>=0.29.0",
    "mistral-common>=1.8.6",
    "peft>=0.14.0",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers>=5.0.0rc0,<6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
